% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichfindCPX.R
\name{enrichfindCPX}
\alias{enrichfindCPX}
\title{Functional Enrichment Analysis for Predicted Modules}
\usage{
enrichfindCPX(
  predcpx,
  threshold = 0.05,
  sources = c("GO", "KEGG", "CORUM", "REAC", "CORUM"),
  p.corrction.method = "bonferroni",
  org = "hsapiens",
  custom_bg = NULL
)
}
\arguments{
\item{predcpx}{A data.frame containing predicted modules resulted from
\code{\link[DeepiCE]{get_clusters}},\code{\link[DeepiCE]{MCL_clustering}},
\code{\link[DeepiCE]{predPPI_MLP}} and \code{\link[DeepiCE]{predPPI_1D.CNN}}.}

\item{threshold}{Custom p-value threshold for significance.}

\item{sources}{A vector of data sources to use.
See \code{\link[gprofiler2]{gost}} for more details.}

\item{p.corrction.method}{The algorithm used for multiple testing
correction;defaults to 'bonferroni'.
See \code{\link[gprofiler2]{gost}} for more details.}

\item{org}{An organism name;defaults to 'hsapiens'.
See \code{\link[gprofiler2]{gost}} for more details.}

\item{custom_bg}{vector of gene names to use as a statistical background.
See \code{\link[gprofiler2]{gost}} for more details.}
}
\value{
A data.frame with the enrichment analysis results.
}
\description{
This function uses \code{\link[gprofiler2]{gost}} function
in \code{gprofiler2} package to perform functional enrichment analysis
for predicted modules.
}
\details{
enrichfindCPX
}
\examples{
# predict complexes
predcpx <- get_clusters()
predcpx <- predcpx[1:10,]
# perform enrichment for KEGG
enrichCPX <-
enrichfindCPX(predcpx,
sources = c("KEGG"))
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
