% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPI.R
\name{getPPI}
\alias{getPPI}
\title{Generating Protein-Protein Interaction (PPI)}
\usage{
getPPI(data)
}
\arguments{
\item{data}{A Co-elution data matrix with proteins in rows and
fractions in columns.}
}
\value{
A concatenated matrix.
}
\description{
This function first removes proteins pairs for which two
proteins never occur in at least one fraction, then concatenates
co-elution profiles for each possible protein pairs to form a master
matrix for input into the learner.
}
\details{
getPPI
}
\examples{
data("HelaCE")
data <- HelaCE[1:100, 1:100]
conc_dat <- getPPI(data)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
