% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCL_clustering.R
\name{MCL_clustering}
\alias{MCL_clustering}
\title{MCL clustering}
\usage{
MCL_clustering(hc_ppi, predcpx, inflation = 4, csize = 2)
}
\arguments{
\item{hc_ppi}{Interactions data containing id1-id2-weight triplets,
generated from \code{\link[DeepiCE]{predPPI_MLP}},
\code{\link[DeepiCE]{MLP_m}},
\code{\link[DeepiCE]{predPPI_1D.CNN}}, and
and \code{\link[DeepiCE]{oneD_CNN}}.}

\item{predcpx}{A data.frame containing predicted modules resulted from
\code{\link[DeepiCE]{get_clusters}}.
\code{\link[DeepiCE]{predPPI_MLP}} and \code{\link[DeepiCE]{predPPI_1D.CNN}}.}

\item{inflation}{MCL inflation parameter. Defaults to 4.}

\item{csize}{Numerical value, the minimum size of the predicted complexes.
Defaults to 2.}
}
\value{
List of refined complexes.
}
\description{
This function applies MCL clustering to further refine
the predicted subnetworks produced by ClusterONE.
}
\details{
MCL_clustering
}
\examples{
# open HC network
hc_ppi <-
read.delim(
system.file("extdata/ppi_input_ClusterOne.txt", package = "DeepiCE"),
header = FALSE)
# predict complexes by ClusterOne
predcpx <-
get_clusters(csize = 3, d = 0.3, p = 2,
mx_overlap = 0.8,
tpath = file.path(system.file("extdata", package = "DeepiCE")))
# MCL clustering
MCL_clusters <- MCL_clustering(hc_ppi, predcpx, inflation = 4, csize = 2)
}
\author{
Matineh Rahmatbakhsh
}
