% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predPPI_MLP.R
\name{predPPI_MLP}
\alias{predPPI_MLP}
\title{Predict Protein-Protein Interactions via MLP Method and
Putative Complexes via a two-stage clustering procedure.}
\usage{
predPPI_MLP(
  data,
  refcpx,
  data_imputed = FALSE,
  bound_left = 2,
  bound_right = 2,
  remove_SinglePeak = FALSE,
  remove_ConsecutivePep = TRUE,
  min_stretch_length = 2,
  scaling = TRUE,
  nlayers = 2,
  powerto1 = 6,
  powerto2 = 7,
  drate = 0.1,
  optimizer = "rmsprop",
  epochs = 50,
  b_size = 128,
  cv_fold = 5,
  plots = FALSE,
  cutoff = 0.5,
  csize = 2,
  d = 0.3,
  p = 2,
  mx_overlap = 0.8,
  inflation = 4,
  tpath = tempdir()
)
}
\arguments{
\item{data}{A data matrix with rows including proteins and fractions
along the columns.}

\item{refcpx}{A list of known reference complexes.}

\item{data_imputed}{If true, detects and imputes missing values. Defaults
to FALSE. See \code{\link{impute_MissingData}}.}

\item{bound_left}{Numeric integer, the minimum number of non-zero
values to the left of a missing value to be replaced with \code{NA}.
Defaults to 2. See \code{\link{impute_MissingData}}.}

\item{bound_right}{Numeric integer, the minimum number of non-zero
values to the right of a missing value to be replaced with \code{NA}.
Defaults to 2. See \code{\link{impute_MissingData}}.}

\item{remove_SinglePeak}{If TRUE, removes Single Peak. Defaults to FALSE.
see \code{\link{RemoveSinglePeak}}.}

\item{remove_ConsecutivePep}{If TRUE, filters peptides by consecutive
peptide detection. Defaults to TRUE.
See \code{\link{filter_ConsecutivePep}}.}

\item{min_stretch_length}{Numeric integer, the minimal length a stretch of
continuous identifications has to have in order not to be removed.
Defaults to 2. See \code{\link{filter_ConsecutivePep}}.}

\item{scaling}{If TRUE, performs row-wise normalization. Defaults to TRUE.
See \code{\link{scaling_m}}.}

\item{nlayers}{Number of hidden layers. Defaults to 2.
See \code{\link{MLP_m}}.}

\item{powerto1}{Integer, the number of neurons in the first hidden layer
as defined by two to the power of this value. Defaults to 6.
See \code{\link{MLP_m}}.}

\item{powerto2}{Integer, the number of neurons in the subsequent hidden
layer as defined by two to the power of this value.Defaults to 7.
See \code{\link{MLP_m}}.}

\item{drate}{Numeric, the dropout rates.
Defaults to 0.1. See \code{\link{MLP_m}}.}

\item{optimizer}{Name of the optimizer.
For most models, this defaults to "rmsprop".See \code{\link{MLP_m}}.}

\item{epochs}{Number of epochs to train the model.
Defaults to 50. See \code{\link{MLP_m}}.}

\item{b_size}{Number of samples per gradient update.
Defaults to 128 See \code{\link{MLP_m}}.}

\item{cv_fold}{Number of partitions for cross-validation; Defaults to 5.
See \code{\link{MLP_m}}.}

\item{plots}{Logical value, indicating whether to plot the performance of
the learning algorithm using k-fold cross-validation; Defaults to FALSE.
These plots are :
\itemize{ \item{pr_plot} - Precision-recall PLOT
\item{roc_plot} - ROC plot
\item{radar_plot} - Radar plot showing
accuracy, F1-score , positive predictive value (PPV), sensitivity (SE)
and MCC.}}

\item{cutoff}{An integer range between [0,1] for specifying
the cutoff for classifier confidence score to subset the high-confidence
interactions.Defaults to 0.5.}

\item{csize}{Numerical value, the minimum size of the predicted complexes.
Defaults to 2. See \code{\link{get_clusters}} or
\code{\link{MCL_clustering}}.}

\item{d}{A number, density of predicted complexes; defaults to 0.3.
See \code{\link{get_clusters}}.}

\item{p}{An integer, penalty value for the inclusion of each node;
defaults to 2. See \code{\link{get_clusters}}.}

\item{mx_overlap}{A number, specifies the maximum allowed
overlap between two clusters; defaults to 0.8.
See \code{\link{get_clusters}}.}

\item{inflation}{MCL inflation parameter. Defaults to 4.
See \code{\link{MCL_clustering}}.}

\item{tpath}{A character string indicating the path to the project
directory. If the directory is
missing, it will be stored in the Temp directory.}
}
\value{
Return three data sets in the project directory including:
 \itemize{
 \item{ppi_input_ClusterOne.txt} - threshold-based network predicted
 from MLP.
 \item{predicted_cpx_clusterOne} - predicted complexes resulted
 from partitioning the threshold-based network predicted from MLP using
 ClusterOne algorithm.
 \item{predicted_cpx_MCL} - predicted complexes resulted
 from partitioning the threshold-based network predicted from MLP using
 MCL algorithm.}
}
\description{
This function first begins by executing several
pre-processing steps to improve the quality of the raw data, followed
by concatenating the co-elution profiles of putative interacting
proteins. Concatenated profiles and class labels generated
from reference complexes are then fed into a multi-layer perceptron (MLP)
deep learning model. These models then generate a weighted protein
interaction network in which edge weights between protein nodes represent
the deep learning model’s probability estimate for interaction.
High-confidence PPIs based on the user-defined threshold is then will be
partitioned via a two-stage approach, clustering first using ClusterOne,
and then further clustering with MCL.
}
\details{
predPPI_MLP
}
\examples{
# Load the input data
data("HelaCE")
HelaCE <- HelaCE[1:500, ]
# Known reference complexes
data("refcpx")
library(tensorflow)
set_random_seed(2)
# Perform prediction
MLP_prediction_outputs <-
predPPI_MLP(HelaCE,
refcpx,
min_stretch_length =3,
cv_fold = 3,
cutoff = 0.5,
epochs = 5,
tpath = tempdir())
}
